Removing App-Cookie Fallback Refactoring WahlGemeindeErgebnis
